scripts/requirements.txt
astroid==2.4.2
autopep8==1.5.3
dt-apriltags==3.1.1
grpcio==1.29.0
grpcio-tools==1.29.0
install==1.3.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.5
picamera==1.13
Pillow==7.1.2
pkg-resources==0.0.0
protobuf==3.12.2
pycodestyle==2.6.0
pylint==2.5.3
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
